LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/write-active-document-expected.txt


FAIL document.write only writes to active documents assert_equals: expected (string) "IFRAME" but got (undefined) undefined
